# Simulation/mock_openapi.yaml
openapi: 3.1.0
info:
  title: Mock API Server
  description: A mock REST API for testing the AWS Data Loader
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/users:
    get:
      summary: Get all users
      description: Retrieve a list of users with optional pagination and search
      operationId: getUsers
      tags:
        - Users
      parameters:
        - name: limit
          in: query
          description: Maximum number of users to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of users to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: search
          in: query
          description: Search term for filtering users by name or email
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  count:
                    type: integer

    post:
      summary: Create a new user
      description: Create a new user with the provided information
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
...
  /api/posts:
    get:
      summary: Get all posts
      description: Retrieve a list of posts with optional filtering
      operationId: getPosts
      tags:
        - Posts
      parameters:
        - name: author_id
          in: query
          description: Filter posts by author ID
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of posts to return
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          description: Number of posts to skip
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  count:
                    type: integer

    post:
      summary: Create a new post
      description: Create a new post with the provided information
      operationId: createPost
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '400':
          description: Bad request
...
  /api/data:
    post:
      summary: Generic data endpoint
      description: A generic endpoint for testing POST requests
      operationId: receiveData
      tags:
        - Testing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Data received successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  received:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
...
components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
          description: Unique identifier for the user
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        age:
          type: integer
          description: User's age
          minimum: 0
    UserInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        age:
          type: integer
          description: User's age
          minimum: 0
    Post:
      type: object
      required:
        - id
        - title
        - content
        - author_id
      properties:
        id:
          type: integer
          description: Unique identifier for the post
        title:
          type: string
          description: Post title
        content:
          type: string
          description: Post content
        author_id:
          type: integer
          description: ID of the author
    PostInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: Post title
        content:
          type: string
          description: Post content
        author_id:
          type: integer
          description: ID of the author
          default: 1
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication (optional for testing)
...